refactor profile fragment
